requirements.txt
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
yaspin>=2.3.0
